Change taskId from int to long.  Generate taskId if not set.

+++ AerospikeClient/Task/ExecuteTask.cs
@@ -21,7 +21,7 @@ namespace Aerospike.Client
 	/// </summary>
 	public sealed class ExecuteTask : Task
 	{
-		private readonly int taskId;
+		private readonly long taskId;
 		private readonly bool scan;
 
 		/// <summary>
